backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-amazon-sp-api==0.26.0
requests==2.31.0
python-dotenv==1.0.0 